+++ config/wdio/wdio.conf.js
@@ -94,6 +94,13 @@ const config = {
 
   axe: {
     inject: true,
+    options: {
+      rules: [
+        // The lowlight theme adheres to a non-default color contrast ratio and fails the default ratio check.
+        // The color-contrast ratio check must be disabled for lowlight theme testing.
+        { id: 'color-contrast', enabled: theme !== 'clinical-lowlight-theme' },
+      ],
+    },
   },
 
   terra: {