+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Disabled color-contrast axe-core rule for clinical-lowlight-theme wdio testing
 
 6.0.0 - (April 21, 2020)
 ----------